reorganized socket event handlers into different files for different categories of event handlers
